Further refactoring of code that creates BSP connections
